refactor: moved formatResponse into functions
