+++ services/core/helpers/index.ts
@@ -2,5 +2,4 @@ export * from "./certificates";
 export * from "./domains";
 export * from "./env-helpers";
 export * from "./env-helpers/environment";
-export * from "./format-response";
 export * from "./lambda";
